Split hallucination into per-format codec components
`Rasterizable` becomes the codec interface rather than a descriptor: it carries
`decode`, `encode` and `sniff`, and each format supplies its own instance from
its own component. The central `RasterBackend` is gone, and with it the core's
knowledge of which formats exist.

Each format chooses its own implementation per platform, through the same
`-jvm`/`-native` source directories galilei and pneumatic.flate use. PNG, JPEG,
GIF and BMP decode through `javax.imageio` on the JVM, whose native codecs
outperform the pure ones, and through their pure codec elsewhere; WebP uses its
pure codec everywhere, since no standard JRE ships a WebP reader. That was the
obstacle to splitting: preferring ImageIO for four formats and the pure codec
for the fifth looked like a core policy, but it was only a policy because core
was choosing. The pure codecs remain in each component's shared directory, so
they are still compiled and still differentially tested against ImageIO.

Recognising a format from its magic bytes now takes explicit candidates:
`Raster(data)` and the `Aggregable` given require a `RasterFormats`.
`hallucination.formats` supplies all five for consumers that want the previous
behaviour in one import. A new codec is a new component with a given.

`Quantization` stays in core, since GIF's palette encoder shares it with JPEG.
tarantula's screenshot decoder now sees only core and png.
