+++ lib/hallucination/src/bmp/hallucination.BmpCodec.scala
