+++ lib/hallucination/src/gif/hallucination.Gif.scala
@@ -32,15 +32,22 @@
                                                                                                   */
 package hallucination
 
+import proscenium.compat.*
+
 import anticipation.*
+import contingency.*
 import gesticulate.*
 
-object Png:
+object Gif:
   def apply(): Rasterizable = rasterization
 
-  given rasterization: Png is Rasterizable:
-    def name: Text = "PNG".tt
-    def mediaType = media"image/png"
+  given rasterization: Gif is Rasterizable:
+    def name: Text = "GIF".tt
+    def mediaType = media"image/gif"
     def alpha: Boolean = true
 
-sealed trait Png
+    def decode(data: Data): Raster raises RasterError = GifBackend.decode(this, data)
+    def encode(raster: Raster): Data = GifBackend.encode(this, raster)
+    def sniff(data: Data): Boolean = (data.length > 2 && data(0) == 0x47 && data(1) == 0x49 && data(2) == 0x46)
+
+sealed trait Gif
