+++ lib/hallucination/src/gif/hallucination.GifCodec.scala
