+++ lib/hallucination/src/gif/hallucination.GifLzw.scala
