+++ lib/hallucination/src/jpeg/hallucination.JpegBitWriter.scala
