+++ lib/hallucination/src/jpeg/hallucination.JpegCodec.scala
