+++ lib/hallucination/src/jpeg/hallucination.JpegEncoder.scala
