+++ lib/hallucination/src/jpeg/hallucination.JpegFdct.scala
