+++ lib/hallucination/src/jpeg/hallucination.JpegHuffman.scala
