+++ lib/hallucination/src/jpeg/hallucination.JpegHuffmanEncoder.scala
