+++ lib/hallucination/src/jpeg/hallucination.JpegParser.scala
