+++ lib/hallucination/src/jpeg/hallucination.JpegReader.scala
