+++ lib/hallucination/src/jpeg/hallucination.JpegUpsampler.scala
