+++ lib/hallucination/src/png/hallucination.PngCodec.scala
