+++ lib/hallucination/src/webp/hallucination.Vp8Bool.scala
