+++ lib/hallucination/src/webp/hallucination.Vp8BoolEncoder.scala
