+++ lib/hallucination/src/webp/hallucination.Vp8Decoder.scala
