+++ lib/hallucination/src/webp/hallucination.Vp8Encoder.scala
