+++ lib/hallucination/src/webp/hallucination.Vp8Filter.scala
