+++ lib/hallucination/src/webp/hallucination.Vp8Predict.scala
