+++ lib/hallucination/src/webp/hallucination.Vp8Tables.scala
