+++ lib/hallucination/src/webp/hallucination.Vp8Transform.scala
