+++ lib/hallucination/src/webp/hallucination.Vp8Yuv.scala
