+++ lib/hallucination/src/webp/hallucination.WebpBitReader.scala
