+++ lib/hallucination/src/webp/hallucination.WebpBitWriter.scala
