+++ lib/hallucination/src/webp/hallucination.WebpCodec.scala
