+++ lib/hallucination/src/webp/hallucination.WebpEncoder.scala
