+++ lib/hallucination/src/webp/hallucination.WebpExtended.scala
