+++ lib/hallucination/src/webp/hallucination.WebpHuffman.scala
