+++ lib/hallucination/src/webp/hallucination.WebpHuffmanEncoder.scala
