+++ lib/hallucination/src/webp/hallucination.WebpLossless.scala
