+++ lib/hallucination/src/webp/hallucination.WebpTransform.scala
