+++ lib/hallucination/src/webp/soundness_hallucination_webp.scala
@@ -30,40 +30,6 @@
 ┃                                                                                                  ┃
 ┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
                                                                                                   */
-package hallucination
+package soundness
 
-import anticipation.*
-import contingency.*
-import fulminate.*
-import vacuous.*
-import proscenium.compat.*
-
-// The pure-Scala backend, selected on Scala.js and WASI. The codecs themselves live in `core`
-// (so they compile, and are differentially tested against `javax.imageio`, on the JVM); this
-// object only dispatches.
-private[hallucination] object RasterBackend:
-  def decode(format: Rasterizable, data: Data): Raster raises RasterError = format.name.s match
-    case "PNG"  => PngCodec.decode(data)
-    case "BMP"  => BmpCodec.decode(data)
-    case "GIF"  => GifCodec.decode(data)
-    case "WEBP" => WebpCodec.decode(data)
-    case "JPEG" => JpegCodec.decode(data)
-    case _      => abort(RasterError(format))
-
-  // Format-agnostic decoding, recognising the format by its opening magic bytes.
-  def decode(data: Data): Raster raises RasterError =
-    if data.length < 4 then abort(RasterError(Unset, RasterError.Reason.Truncated))
-    else if (data(0)&0xff) == 0x89 && data(1) == 0x50 then PngCodec.decode(data)
-    else if data(0) == 0x47 && data(1) == 0x49 && data(2) == 0x46 then GifCodec.decode(data)
-    else if data(0) == 0x42 && data(1) == 0x4d then BmpCodec.decode(data)
-    else if JpegCodec.isJpeg(data) then JpegCodec.decode(data)
-    else if WebpCodec.isWebp(data) then WebpCodec.decode(data)
-    else abort(RasterError(Unset))
-
-  def encode(format: Rasterizable, raster: Raster): Data = format.name.s match
-    case "PNG"  => PngCodec.encode(raster)
-    case "BMP"  => BmpCodec.encode(raster)
-    case "GIF"  => GifCodec.encode(raster)
-    case "WEBP" => WebpCodec.encode(raster)
-    case "JPEG" => JpegEncoder.encode(raster)
-    case _      => panic(m"the ${format.name} format has no native encoder")
+export hallucination.Webp
